// src/app/core/models/registrationEvent.ts
export interface RegistrationEvent{
     eventName: string;
     placeName: string;
     peopleCapacityOfEvent: string;
     priceOfBasieTicket: string;
     priceOfDeluxeTicket: string;
     priceOfPremiumTicket: string;
     locationLatitude: number;
     locationLongitude: number;
     eventSetup: string;
}